added model types in frontend
